add property testing
