+++ helix-core/Cargo.toml
@@ -45,4 +45,6 @@ etcetera = "0.4"
 textwrap = "0.15.1"
 
 [dev-dependencies]
-quickcheck = { version = "1", default-features = false }
+
+proptest = {version = "1.0", default-features=false, features = ["std"]}
+similar-asserts = "1.4.2"
