+++ helix-core/src/syntax.rs
@@ -2209,633 +2209,4 @@ impl<I: Iterator<Item = HighlightEvent>, R: Iterator<Item = HighlightEvent>> Ite
 }
 
 #[cfg(test)]
-mod test {
-    use super::*;
-    use crate::{Rope, Transaction};
-
-    #[test]
-    fn test_textobject_queries() {
-        let query_str = r#"
-        (line_comment)+ @quantified_nodes
-        ((line_comment)+) @quantified_nodes_grouped
-        ((line_comment) (line_comment)) @multiple_nodes_grouped
-        "#;
-        let source = Rope::from_str(
-            r#"
-/// a comment on
-/// multiple lines
-        "#,
-        );
-
-        let loader = Loader::new(Configuration { language: vec![] });
-        let language = get_language("Rust").unwrap();
-
-        let query = Query::new(language, query_str).unwrap();
-        let textobject = TextObjectQuery { query };
-        let mut cursor = QueryCursor::new();
-
-        let config = HighlightConfiguration::new(language, "", "", "").unwrap();
-        let syntax = Syntax::new(&source, Arc::new(config), Arc::new(loader));
-
-        let root = syntax.tree().root_node();
-        let mut test = |capture, range| {
-            let matches: Vec<_> = textobject
-                .capture_nodes(capture, root, source.slice(..), &mut cursor)
-                .unwrap()
-                .collect();
-
-            assert_eq!(
-                matches[0].byte_range(),
-                range,
-                "@{} expected {:?}",
-                capture,
-                range
-            )
-        };
-
-        test("quantified_nodes", 1..36);
-        // NOTE: Enable after implementing proper node group capturing
-        // test("quantified_nodes_grouped", 1..36);
-        // test("multiple_nodes_grouped", 1..36);
-    }
-
-    #[test]
-    fn test_parser() {
-        let highlight_names: Vec<String> = [
-            "attribute",
-            "constant",
-            "function.builtin",
-            "function",
-            "keyword",
-            "operator",
-            "property",
-            "punctuation",
-            "punctuation.bracket",
-            "punctuation.delimiter",
-            "string",
-            "string.special",
-            "tag",
-            "type",
-            "type.builtin",
-            "variable",
-            "variable.builtin",
-            "variable.parameter",
-        ]
-        .iter()
-        .cloned()
-        .map(String::from)
-        .collect();
-
-        let loader = Loader::new(Configuration { language: vec![] });
-
-        let language = get_language("Rust").unwrap();
-        let config = HighlightConfiguration::new(
-            language,
-            &std::fs::read_to_string("../runtime/grammars/sources/rust/queries/highlights.scm")
-                .unwrap(),
-            &std::fs::read_to_string("../runtime/grammars/sources/rust/queries/injections.scm")
-                .unwrap(),
-            "", // locals.scm
-        )
-        .unwrap();
-        config.configure(&highlight_names);
-
-        let source = Rope::from_str(
-            "
-            struct Stuff {}
-            fn main() {}
-        ",
-        );
-        let syntax = Syntax::new(&source, Arc::new(config), Arc::new(loader));
-        let tree = syntax.tree();
-        let root = tree.root_node();
-        assert_eq!(root.kind(), "source_file");
-
-        assert_eq!(
-            root.to_sexp(),
-            concat!(
-                "(source_file ",
-                "(struct_item name: (type_identifier) body: (field_declaration_list)) ",
-                "(function_item name: (identifier) parameters: (parameters) body: (block)))"
-            )
-        );
-
-        let struct_node = root.child(0).unwrap();
-        assert_eq!(struct_node.kind(), "struct_item");
-    }
-
-    #[test]
-    fn test_input_edits() {
-        use tree_sitter::InputEdit;
-
-        let doc = Rope::from("hello world!\ntest 123");
-        let transaction = Transaction::change(
-            &doc,
-            vec![(6, 11, Some("test".into())), (12, 17, None)].into_iter(),
-        );
-        let edits = generate_edits(&doc, transaction.changes());
-        // transaction.apply(&mut state);
-
-        assert_eq!(
-            edits,
-            &[
-                InputEdit {
-                    start_byte: 6,
-                    old_end_byte: 11,
-                    new_end_byte: 10,
-                    start_position: Point { row: 0, column: 6 },
-                    old_end_position: Point { row: 0, column: 11 },
-                    new_end_position: Point { row: 0, column: 10 }
-                },
-                InputEdit {
-                    start_byte: 12,
-                    old_end_byte: 17,
-                    new_end_byte: 12,
-                    start_position: Point { row: 0, column: 12 },
-                    old_end_position: Point { row: 1, column: 4 },
-                    new_end_position: Point { row: 0, column: 12 }
-                }
-            ]
-        );
-
-        // Testing with the official example from tree-sitter
-        let mut doc = Rope::from("fn test() {}");
-        let transaction =
-            Transaction::change(&doc, vec![(8, 8, Some("a: u32".into()))].into_iter());
-        let edits = generate_edits(&doc, transaction.changes());
-        transaction.apply(&mut doc);
-
-        assert_eq!(doc, "fn test(a: u32) {}");
-        assert_eq!(
-            edits,
-            &[InputEdit {
-                start_byte: 8,
-                old_end_byte: 8,
-                new_end_byte: 14,
-                start_position: Point { row: 0, column: 8 },
-                old_end_position: Point { row: 0, column: 8 },
-                new_end_position: Point { row: 0, column: 14 }
-            }]
-        );
-    }
-
-    #[test]
-    fn test_load_runtime_file() {
-        // Test to make sure we can load some data from the runtime directory.
-        let contents = load_runtime_file("rust", "indents.scm").unwrap();
-        assert!(!contents.is_empty());
-
-        let results = load_runtime_file("rust", "does-not-exist");
-        assert!(results.is_err());
-    }
-
-    #[test]
-    fn test_sample_highlight_event_stream_merge() {
-        use HighlightEvent::*;
-
-        /*
-        Left:
-                              2          3
-                       |-----------|-----------|
-
-                                    1
-            |-----------------------------------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9  10  11  12
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 0, end: 3 },
-            HighlightStart(Highlight(2)),
-            Source { start: 3, end: 6 },
-            HighlightEnd, // ends 2
-            HighlightStart(Highlight(3)),
-            Source { start: 6, end: 9 },
-            HighlightEnd, // ends 3
-            Source { start: 9, end: 12 },
-            HighlightEnd, // ends 1
-        ]
-        .into_iter();
-
-        /*
-        Right:
-                       100            200
-                    |-------|   |---------------|
-
-            |---|---|---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9  10  11  12
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(100)),
-                Source { start: 2, end: 4 },
-                HighlightEnd, // ends 100
-                HighlightStart(Highlight(200)),
-                Source { start: 5, end: 9 },
-                HighlightEnd, // ends 200
-            ]
-            .into_iter(),
-        );
-
-        /*
-        Output:
-                         100          200
-                        |---|   |---------------|
-
-                     100      2          3
-                    |---|-----------|-----------|
-
-                                    1
-            |-----------------------------------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9  10  11  12
-        */
-        let output: Vec<_> = merge(left, right).collect();
-
-        assert_eq!(
-            output,
-            &[
-                HighlightStart(Highlight(1)),
-                Source { start: 0, end: 2 },
-                HighlightStart(Highlight(100)),
-                Source { start: 2, end: 3 },
-                HighlightEnd, // ends 100
-                HighlightStart(Highlight(2)),
-                HighlightStart(Highlight(100)),
-                Source { start: 3, end: 4 },
-                HighlightEnd, // ends 100
-                Source { start: 4, end: 5 },
-                HighlightStart(Highlight(200)),
-                Source { start: 5, end: 6 },
-                HighlightEnd, // ends 200
-                HighlightEnd, // ends 2
-                HighlightStart(Highlight(3)),
-                HighlightStart(Highlight(200)),
-                Source { start: 6, end: 9 },
-                HighlightEnd, // ends 200
-                HighlightEnd, // ends 3
-                Source { start: 9, end: 12 },
-                HighlightEnd // ends 1
-            ],
-        );
-    }
-
-    #[test]
-    fn test_highlight_event_stream_merge_overlapping() {
-        use HighlightEvent::*;
-
-        /*
-        Left:
-                   1, 2, 3
-            |-------------------|
-
-            |---|---|---|---|---|
-            0   1   2   3   4   5
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            HighlightStart(Highlight(2)),
-            HighlightStart(Highlight(3)),
-            Source { start: 0, end: 5 },
-            HighlightEnd, // ends 3
-            HighlightEnd, // ends 2
-            HighlightEnd, // ends 1
-        ]
-        .into_iter();
-
-        /*
-        Right:
-                      4
-            |-------------------|
-
-            |---|---|---|---|---|
-            0   1   2   3   4   5
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(4)),
-                Source { start: 0, end: 5 },
-                HighlightEnd, // ends 4
-            ]
-            .into_iter(),
-        );
-
-        /*
-        Output:
-                  1, 2, 3, 4
-            |-------------------|
-
-            |---|---|---|---|---|
-            0   1   2   3   4   5
-        */
-        let output: Vec<_> = merge(left, right).collect();
-
-        assert_eq!(
-            output,
-            &[
-                HighlightStart(Highlight(1)),
-                HighlightStart(Highlight(2)),
-                HighlightStart(Highlight(3)),
-                HighlightStart(Highlight(4)),
-                Source { start: 0, end: 5 },
-                HighlightEnd, // ends 4
-                HighlightEnd, // ends 3
-                HighlightEnd, // ends 2
-                HighlightEnd, // ends 1
-            ],
-        );
-    }
-
-    #[test]
-    fn test_highlight_event_stream_merge_right_is_truncated() {
-        use HighlightEvent::*;
-        // This can happen when there are selections outside of the
-        // viewport. `left` is the syntax highlight event stream and
-        // `right` is the `span_iter` of selections/cursors.
-
-        /*
-        Left:
-                              1
-                    |-------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 2, end: 7 },
-            HighlightEnd, // ends 1
-        ]
-        .into_iter();
-
-        /*
-        Right:
-              2                 3                 4
-            |---|-------------------------------|---|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(2)),
-                Source { start: 0, end: 1 },
-                HighlightEnd, // ends 2
-                HighlightStart(Highlight(3)),
-                Source { start: 1, end: 9 },
-                HighlightEnd, // ends 3
-                HighlightStart(Highlight(4)),
-                Source { start: 9, end: 10 },
-                HighlightEnd, // ends 4
-            ]
-            .into_iter(),
-        );
-
-        // 2 and 4 are out of range and are discarded. 3 is truncated at the
-        // beginning but allowed to finish after the `left`. This is a special
-        // case for the trailing space from selection highlights.
-        /*
-        Output:
-                             1, 3           3
-                    |-------------------|-------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let output: Vec<_> = merge(left, right).collect();
-
-        assert_eq!(
-            output,
-            &[
-                HighlightStart(Highlight(1)),
-                HighlightStart(Highlight(3)),
-                Source { start: 2, end: 7 },
-                HighlightEnd, // ends 3
-                HighlightEnd, // ends 1
-                HighlightStart(Highlight(3)),
-                Source { start: 7, end: 9 },
-                HighlightEnd, // ends 3
-            ],
-        );
-    }
-
-    #[test]
-    fn test_highlight_event_stream_right_ends_before_left_starts() {
-        use HighlightEvent::*;
-
-        /*
-        Left:
-                                          1
-                                |-------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 5, end: 10 },
-            HighlightEnd, // ends 1
-        ];
-
-        /*
-        Right:
-                    2
-            |---------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(2)),
-                Source { start: 0, end: 4 },
-                HighlightEnd, // ends 2
-            ]
-            .into_iter(),
-        );
-
-        // Left starts after right ends. Right is discarded.
-        let output: Vec<_> = merge(left.clone().into_iter(), right).collect();
-        assert_eq!(output, left);
-    }
-
-    #[test]
-    fn test_highlight_event_stream_right_ends_as_left_starts() {
-        use HighlightEvent::*;
-
-        /*
-        Left:
-                                          1
-                                |-------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 5, end: 10 },
-            HighlightEnd, // ends 1
-        ];
-
-        /*
-        Right:
-                      2
-            |-------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(2)),
-                Source { start: 0, end: 5 },
-                HighlightEnd, // ends 2
-            ]
-            .into_iter(),
-        );
-
-        // Right is discarded if the range ends in the same place as left starts.
-        let output: Vec<_> = merge(left.clone().into_iter(), right).collect();
-        assert_eq!(output, left);
-    }
-
-    #[test]
-    fn test_highlight_event_stream_merge_layered_overlapping() {
-        use HighlightEvent::*;
-
-        /*
-        Left:
-                              1
-                        |-----------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 3, end: 6 },
-            HighlightEnd, // ends 1
-        ]
-        .into_iter();
-
-        /*
-        Right:
-                              3     4 (0-width)
-                        |-----------|
-
-                                2
-                |-------------------------------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(2)),
-                Source { start: 1, end: 3 },
-                HighlightStart(Highlight(3)),
-                Source { start: 3, end: 6 },
-                HighlightEnd, // ends 3
-                HighlightStart(Highlight(4)),
-                // Trimmed zero-width Source.
-                HighlightEnd, // ends 4
-                Source { start: 6, end: 9 },
-                HighlightEnd, // ends 2
-            ]
-            .into_iter(),
-        );
-
-        /*
-        Output:
-                            1,2,3
-                        |-----------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let output: Vec<_> = merge(left, right).collect();
-
-        assert_eq!(
-            output,
-            &[
-                HighlightStart(Highlight(1)),
-                HighlightStart(Highlight(2)),
-                HighlightStart(Highlight(3)),
-                Source { start: 3, end: 6 },
-                HighlightEnd, // ends 3
-                HighlightEnd, // ends 2
-                HighlightEnd, // ends 1
-            ],
-        );
-    }
-
-    #[test]
-    fn test_highlight_event_stream_merge_double_zero_width_span() {
-        use HighlightEvent::*;
-        // This is possible when merging two syntax highlight iterators.
-        // Syntax highlight iterators may produce a HighlightStart
-        // immediately followed by a HighlightEnd.
-
-        /*
-        Left:
-                              1     2 (0-width)
-                        |-----------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let left = vec![
-            HighlightStart(Highlight(1)),
-            Source { start: 3, end: 6 },
-            HighlightEnd, // ends 1
-            HighlightStart(Highlight(2)),
-            HighlightEnd, // ends 2
-        ]
-        .into_iter();
-
-        /*
-        Right:
-                              3     4 (0-width)
-                        |-----------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let right = Box::new(
-            vec![
-                HighlightStart(Highlight(3)),
-                Source { start: 3, end: 6 },
-                HighlightEnd, // ends 3
-                HighlightStart(Highlight(4)),
-                HighlightEnd, // ends 4
-            ]
-            .into_iter(),
-        );
-
-        /*
-        Output:
-                             1,3
-                        |-----------|
-
-            |---|---|---|---|---|---|---|---|---|---|
-            0   1   2   3   4   5   6   7   8   9   10
-        */
-        let output: Vec<_> = merge(left, right).collect();
-
-        assert_eq!(
-            output,
-            &[
-                HighlightStart(Highlight(1)),
-                HighlightStart(Highlight(3)),
-                Source { start: 3, end: 6 },
-                HighlightEnd, // ends 3
-                HighlightEnd, // ends 1
-                HighlightStart(Highlight(4)),
-                // Zero-width Source span.
-                HighlightEnd, // ends 4
-            ],
-        );
-    }
-}
+mod test;
